Refactor PdfEditor into reusable hooks and overlay renderer
- Extract PDF loading and thumbnail generation into usePdfDocument hook
- Move preview overlay drawing into utils/overlayRenderer
- Move page-order remapping helpers into utils/pageOrderUtils
- Deduplicate header/footer drawing logic in pdfEditOperations
